SelfDrivingCar/requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
greenlet==0.4.15
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pylint==2.4.4
pypiwin32==223
python-engineio==3.11.1
python-socketio==4.4.0
pywin32==227
six==1.13.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
